refactor(acquisition_maximizer): fix typo
